添加根据类型查询分类接口(同步视频)

+++ sky-take-out/sky-server/src/main/java/com/sky/service/impl/CategoryServiceImpl.java
@@ -5,12 +5,16 @@ package com.sky.service.impl;
 
 import com.github.pagehelper.Page;
 import com.github.pagehelper.PageHelper;
+import com.sky.constant.MessageConstant;
 import com.sky.constant.StatusConstant;
 import com.sky.context.BaseContext;
 import com.sky.dto.CategoryDTO;
 import com.sky.dto.CategoryPageQueryDTO;
 import com.sky.entity.Category;
+import com.sky.exception.DeletionNotAllowedException;
 import com.sky.mapper.CategoryMapper;
+import com.sky.mapper.DishMapper;
+import com.sky.mapper.SetmealMapper;
 import com.sky.result.PageResult;
 import com.sky.service.CategoryService;
 import org.springframework.beans.BeanUtils;
@@ -25,6 +29,10 @@ public class CategoryServiceImpl implements CategoryService {
 
     @Autowired
     private CategoryMapper categoryMapper;
+    @Autowired
+    private DishMapper dishMapper;
+    @Autowired
+    private SetmealMapper setmealMapper;
 
     /**分类分页查询
      *
@@ -47,7 +55,21 @@ public class CategoryServiceImpl implements CategoryService {
      * @param id
      * @return
      */
-    public void delete(Long id) {
+    public void deleteById(Long id) {
+//        查询分类是否关联菜品 如果关联就抛出异常
+        Integer count = dishMapper.countByCategoryId(id);
+        if (count > 0) {
+//            当前分类有菜品不能删除
+            throw new DeletionNotAllowedException(MessageConstant.CATEGORY_BE_RELATED_BY_DISH);
+        }
+//        查询分类是否关联套餐 如果关联就抛出异常
+        count = setmealMapper.countByCategoryId(id);
+        if (count > 0) {
+//            当前分类有套餐 不能删除
+            throw new DeletionNotAllowedException(MessageConstant.CATEGORY_BE_RELATED_BY_SETMEAL);
+        }
+
+
         categoryMapper.delete(id);
     }
 
@@ -99,8 +121,8 @@ public class CategoryServiceImpl implements CategoryService {
      * @param type
      * @return
      */
-    public void list(long type) {
-
+    public List<Category> list(Integer type) {
+        return categoryMapper.list(type);
     }
 
 }
